Combine search-utils and search-filter
