+++ src/app/search/loadouts/loadout-search-filter.ts
@@ -8,8 +8,7 @@ import { loadoutsSelector } from 'app/loadout-drawer/loadouts-selector';
 import { LoadoutsByItem, loadoutsByItemSelector } from 'app/loadout-drawer/selectors';
 import { d2ManifestSelector } from 'app/manifest/selectors';
 import { buildFiltersMap, buildSearchConfig } from 'app/search/search-config';
-import { makeSearchFilterFactory } from 'app/search/search-filter';
-import { parseAndValidateQuery } from 'app/search/search-utils';
+import { makeSearchFilterFactory, parseAndValidateQuery } from 'app/search/search-filter';
 import memoizeOne from 'memoize-one';
 import { createSelector } from 'reselect';
 import { LoadoutFilterContext, LoadoutSuggestionsContext } from './loadout-filter-types';
